requirements.txt
aenum==3.1.11
alabaster==0.7.12
appdirs==1.4.4
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0git
arrow==1.2.2
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.5
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backend==0.2.4.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0c
bs4==0.0.1
cachetools==5.2.0
case-conversion==2.1.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
clipboard==0.0.4
colorama==0.4.5
colour==0.1.5
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.2
crispy-bootstrap5==0.6
cryptography==37.0.4
cssselect==1.1.0
cssutils==2.5.1
cursor==1.3.5
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.5
dj-database-url==1.0.0
Django==3.2.15
django-allauth==0.51.0
django-cors-headers==3.13.0
django-coverage-plugin==2.0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-extensions==3.2.0
django-filter==22.1
django-model-utils==4.2.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-rest-authemail==2.1.4
django-rest-authtoken==2.1.4
django-sslserver==0.22
django-stubs==1.12.0
django-stubs-ext==0.5.0
django-templated-mail==1.1.1
django-uwsgi==0.2.2
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.19
emails==0.6
executing==0.9.1
factory-boy==3.2.1
Faker==13.15.1
filelock==3.7.1
flake8==5.0.4
flake8-isort==4.2.0
fuzzywuzzy==0.18.0
gender-guesser==0.4.0
gmailapi==0.3.1
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
hiredis==2.0.0
html-text==0.5.2
httplib2==0.20.4
identify==2.5.3
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
imagesize==1.4.1
infi.systray==0.1.12
inflect==6.0.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
lazy-object-proxy==1.7.1
livereload==2.6.3
lxml==4.9.1
mail-parser==3.15.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
maybe-else==0.2.1
mbstrdecoder==1.1.1
mccabe==0.7.0
moviepy==1.0.3
msoffcrypto-tool==5.0.0
multipledispatch==0.6.0
mypy==0.971
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.3
parsedatetime==2.6
parso==0.8.3
pathmagic==0.3.14
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
premailer==3.10.0
prettierfier==1.0.3
proglog==0.1.10
prompt-toolkit==3.0.30
protobuf==4.21.5
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
pydub==0.25.1
pyflakes==2.5.0
Pygments==2.12.0
pyinstrument==4.2.0
pyiotools==0.3.18
PyJWT==2.4.0
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pymiscutils==0.3.14
pyparsing==3.0.9
PyPDF2==2.10.0
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pysubtypes==0.3.18
pytest==7.1.2
pytest-django==4.5.2
pytest-sugar==0.9.5
python-dateutil==2.8.2
python-docx==0.8.11
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
readchar==4.0.0
redis==4.3.4
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.3.0
tabulate==0.8.10
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typepy==1.3.0
types-pytz==2022.1.2
types-PyYAML==6.0.11
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.3
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.1
whitenoise==6.2.0
wrapt==1.14.1
XlsxWriter==3.0.3
